requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-MySQL==1.5.2
gunicorn==21.2.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.2
protobuf==4.21.12
PyMySQL==1.1.0
python-dotenv==1.0.0
waitress==2.1.2
Werkzeug==2.3.7